add some modle tag

+++ internal/model/ip_info.go
@@ -4,14 +4,14 @@ import "gorm.io/gorm"
 
 type IpInfo struct {
 	ID      uint64 `gorm:"primaryKey"`
-	IpAddr  string
-	Count   int
+	IpAddr  string //  唯一  index
+	Count   uint64
 	Address string
 	ApiList []Api
 }
 type Api struct {
 	gorm.Model
 	Router     string
-	Count      int
+	Count      uint64
 	IpInfoList []IpInfo
 }
